refactor(link): folder structure
