+++ packages/asana-link/action.yml
@@ -14,17 +14,20 @@ inputs:
     description: 'Prefix that goes before the asana link comment'
     required: false
   on_open_action:
-    description: 'The action to perform when a PR is opened (CLOSE|MOVE <Section Name>)'
+    description:
+      'The action to perform when a PR is opened (CLOSE|MOVE <Section Name>)'
     required: false
   fail_on_no_task:
     description: 'Make the action fail if there are no found tasks'
     required: false
   on_merge_action:
-    description: 'The action to perform when a PR is merged (CLOSE|MOVE <Section Name>).  Defaults to CLOSE'
+    description:
+      'The action to perform when a PR is merged (CLOSE|MOVE <Section
+      Name>).  Defaults to CLOSE'
     required: false
 branding:
-  icon: 'chevron-right'  
+  icon: 'chevron-right'
   color: 'gray-dark'
 runs:
   using: 'node12'
-  main: 'index.js'
+  main: 'src/index.js'
